Project/requirements.txt
Django==1.9.7
django-appconf==1.0.2
django-extensions==1.7.4
django-geoposition==0.3.0
django-messages==0.5.3
django-user-accounts==1.3.1
djangorestframework==3.5.1
geopy==1.11.0
oauthlib==2.0.0
Pillow==3.3.1
pkg-resources==0.0.0
pydotplus==2.0.2
PyJWT==1.4.2
pyparsing==2.1.9
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
